+++ src/client/utils/index.ts
@@ -1,2 +1,3 @@
 export { request, fullUrl, RequestData, RequestMethod } from './request';
 export { toTime } from './time';
+export { areDifferent } from './compare';