+++ src/server/management/injectors.ts
@@ -23,6 +23,9 @@ export function saveInjectorSettings(server: KrasServer) {
 
     if (injector) {
       injector.setOptions(options);
+      return res.sendStatus(200);
     }
+
+    res.sendStatus(404);
   };
 }